version: move to module, avoid circular imports
