+++ lib/spack/spack/version/__init__.py
@@ -1,0 +1,58 @@
+# Copyright 2013-2023 Lawrence Livermore National Security, LLC and other
+# Spack Project Developers. See the top-level COPYRIGHT file for details.
+#
+# SPDX-License-Identifier: (Apache-2.0 OR MIT)
+
+"""
+This module implements Version and version-ish objects.  These are:
+
+StandardVersion: A single version of a package.
+ClosedOpenRange: A range of versions of a package.
+VersionList: A ordered list of Version and VersionRange elements.
+
+The set of Version and ClosedOpenRange is totally ordered wiht <
+defined as Version(x) < VersionRange(Version(y), Version(x))
+if Version(x) <= Version(y).
+"""
+
+from .common import (
+    VersionChecksumError,
+    VersionError,
+    VersionLookupError,
+    infinity_versions,
+    is_git_version,
+)
+from .version_types import (
+    ClosedOpenRange,
+    GitVersion,
+    StandardVersion,
+    Version,
+    VersionList,
+    VersionRange,
+    from_string,
+    next_version,
+    prev_version,
+    ver,
+)
+
+#: This version contains all possible versions.
+any_version: VersionList = VersionList([":"])
+
+__all__ = [
+    "Version",
+    "VersionRange",
+    "ver",
+    "from_string",
+    "is_git_version",
+    "infinity_versions",
+    "prev_version",
+    "next_version",
+    "VersionList",
+    "ClosedOpenRange",
+    "StandardVersion",
+    "GitVersion",
+    "VersionError",
+    "VersionChecksumError",
+    "VersionLookupError",
+    "any_version",
+]
